put fake kernels and fake traversal functions in test.h for unit test
