+++ 3dp/tree.cxx
@@ -1,21 +1,7 @@
+#include <cassert>
 #include "build_tree.h"
+#include "test.h"
 using namespace exafmm;
-
-//! Recursive call to post-order tree traversal for evaluating monopoles
-void evalMonopole(Cell * Ci) {
-  for (Cell * Cj=Ci->CHILD; Cj!=Ci->CHILD+Ci->NCHILD; ++Cj) {   // Loop over child cells
-    evalMonopole(Cj);                                           //  Recursive call for child cell
-  }                                                             // End loop over child cells
-  Ci->M.resize(1, 0.0);                                         // Allocate and initialize multipole coefs
-  if (Ci->NCHILD==0) {                                          // If Ci is leaf
-    for (Body * B=Ci->BODY; B!=Ci->BODY+Ci->NBODY; ++B)         //  Loop over bodies in Ci
-      Ci->M[0] += B->q;                                         //   Calculate monopole
-  }                                                             // End if
-  else {                                                        // If Ci is not leaf
-    for (Cell * Cj=Ci->CHILD; Cj!=Ci->CHILD+Ci->NCHILD; ++Cj)   //  Loop over Ci's child cells
-      Ci->M[0] += Cj->M[0];                                     //   Calculate monopole
-  }                                                             // End else
-}
 
 int main(int argc, char ** argv) {
   const int numBodies = atoi(argv[1]);                          // Number of bodies
@@ -37,9 +23,12 @@ int main(int argc, char ** argv) {
   Cells cells = buildTree(bodies);                              // Build tree
 
   //! Evaluate monopoles (Upward Pass)
-  evalMonopole(&cells[0]);
+  test::upwardPass(&cells[0]);
   
+  //! Check Answer
   printf("%-20s : %i\n", "Num of Bodies", numBodies);
-  printf("%-20s : %f \n", "Root's Monopole", cells[0].M[0]);
+  printf("--- %-18s ------------\n", "Checking monopole");
+  assert(numBodies == std::real(cells[0].M[0]));
+  printf("--- %-18s ------------\n", "Assertion Passed!");
   return 0;
 }
